Post rebase fixes.

+++ src/include/simeng/pipeline/CapacityAwarePortAllocator.hh
@@ -31,7 +31,7 @@ class CapacityAwarePortAllocator : public PortAllocator {
   void deallocate(uint16_t port, const uint16_t stallCycles = 0) override;
 
   void setRSSizeGetter(
-      std::function<void(std::vector<uint64_t>&)> rsSizes) override;
+      std::function<void(std::vector<uint32_t>&)> rsSizes) override;
 
   /** Tick the port allocator to allow it to process internal tasks. */
   void tick() override;
@@ -47,7 +47,7 @@ class CapacityAwarePortAllocator : public PortAllocator {
 
   std::vector<std::map<uint16_t, uint8_t>> rsPortMappings_;
 
-  std::function<void(std::vector<uint64_t>&)> rsSizes_;
+  std::function<void(std::vector<uint32_t>&)> rsSizes_;
 };
 
 }  // namespace pipeline
